# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevSpaces Workspace API",
    "description": "API for managing DevSpaces workspaces (IntelliJ and VS Code)",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "description": "Root endpoint with API information",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Root  Get"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Health check endpoint",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/workspaces/{namespace}": {
      "get": {
        "summary": "List Workspaces",
        "description": "List all DevWorkspaces in a namespace",
        "operationId": "list_workspaces_workspaces__namespace__get",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/intellij": {
      "post": {
        "summary": "Create Intellij Workspace",
        "description": "Create an IntelliJ DevWorkspace",
        "operationId": "create_intellij_workspace_workspaces_intellij_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelliJWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/vscode": {
      "post": {
        "summary": "Create Vscode Workspace",
        "description": "Create a VS Code DevWorkspace",
        "operationId": "create_vscode_workspace_workspaces_vscode_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VSCodeWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{namespace}/{name}": {
      "get": {
        "summary": "Get Workspace",
        "description": "Get a specific DevWorkspace",
        "operationId": "get_workspace_workspaces__namespace___name__get",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Workspace",
        "description": "Delete a DevWorkspace",
        "operationId": "delete_workspace_workspaces__namespace___name__delete",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DeleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "DeleteResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntelliJWorkspaceRequest": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace",
            "description": "OpenShift namespace",
            "default": "admin-devspaces"
          },
          "git_repo": {
            "type": "string",
            "title": "Git Repo",
            "description": "Git repository URL",
            "default": "https://github.com/spring-projects/spring-petclinic"
          },
          "memory_limit": {
            "type": "string",
            "title": "Memory Limit",
            "description": "Memory limit",
            "default": "8Gi"
          },
          "cpu_limit": {
            "type": "string",
            "title": "Cpu Limit",
            "description": "CPU limit",
            "default": "2000m"
          }
        },
        "type": "object",
        "title": "IntelliJWorkspaceRequest"
      },
      "VSCodeWorkspaceRequest": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace",
            "description": "OpenShift namespace",
            "default": "admin-devspaces"
          },
          "git_repo": {
            "type": "string",
            "title": "Git Repo",
            "description": "Git repository URL",
            "default": "https://github.com/quarkusio/quarkus-quickstarts"
          },
          "memory_limit": {
            "type": "string",
            "title": "Memory Limit",
            "description": "Memory limit",
            "default": "4Gi"
          },
          "cpu_limit": {
            "type": "string",
            "title": "Cpu Limit",
            "description": "CPU limit",
            "default": "2000m"
          }
        },
        "type": "object",
        "title": "VSCodeWorkspaceRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceListResponse": {
        "properties": {
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "type": "array",
            "title": "Workspaces"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "workspaces",
          "total"
        ],
        "title": "WorkspaceListResponse"
      },
      "WorkspaceResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "namespace",
          "phase"
        ],
        "title": "WorkspaceResponse"
      }
    }
  }
}